+++ code/main.R
@@ -1,5 +1,4 @@
 #!/usr/bin/env Rscript
-rlang::global_entrace()
 library(argparse)
 library(glue)
 library(MOSuite)
@@ -14,8 +13,8 @@ setup_capsule_environment()
 parser <- ArgumentParser(description = "Create multi-omic dataset from files")
 
 parser$add_argument("--delim", type="character", required=FALSE, default=",")
-parser$add_argument("--regex_count", type="character", default="count.*.[ct]sv*")
-parser$add_argument("--regex_sample", type="character", default="sample.*.[ct]sv*")
+parser$add_argument("--regex_count", type="character", default="count.*\\.[ct]sv(\\.gz)?")
+parser$add_argument("--regex_sample", type="character", default="sample.*\\.[ct]sv(\\.gz)?")
 
 args <- parser$parse_args()
 
@@ -31,11 +30,14 @@ sample_meta_files <- Filter(\(x) str_detect(x, regex(regex_sample, ignore_case =
 # validate inputs
 if (length(count_files) == 0) {
     stop(glue("No counts data file was found. Please place a csv or tsv file in data/ that matches this regex: {regex_count}"))
+} else if (length(count_files) > 1) {
+    warning(glue("Multiple counts data files were found matching {regex_count}: {paste(count_files, collapse=', ')}"))
 }
 if (length(sample_meta_files) == 0) {
     stop(glue("No sample metadata file was found. Please place a csv or tsv file in data/ that matches this regex: {regex_sample}"))
+} else if (length(sample_meta_files) > 1) {
+    warning(glue("Multiple sample metadata files were found matching {regex_sample}: {paste(sample_meta_files, collapse=', ')}"))
 }
-
 count_filename <- count_files[1]
 sample_metadata_filename <- sample_meta_files[1]
 message(glue("Counts data file: {count_filename}"))
@@ -47,4 +49,4 @@ moo <- create_multiOmicDataSet_from_files(
     feature_counts_filepath = count_filename,
     delim = args$delim
 )
-write_rds(moo, file.path(getOption("moo_plots_dir"), "..", "moo", "moo.rds"))
+write_rds(moo, file.path(getOption("moo_plots_dir"), "..", "moo", "moo-raw.rds"))